almost done

+++ utilities/utilities.js
@@ -1,11 +1,11 @@
 function getInputValueById(id){
     const ammout = document.getElementById(id).value;
-    const convertedAmmount = parseInt(ammout);
+    const convertedAmmount = parseFloat(ammout);
     return convertedAmmount;
 };
 function getInnerTextById(id){
     const ammout = document.getElementById(id).innerText;
-    const convertedAmmount = parseInt(ammout);
+    const convertedAmmount = parseFloat(ammout);
     return convertedAmmount;
 };
 function setInnerTextByIdAndValue(id, value){
